pyproject.toml
[tool.black]
line-length = 120

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
skip_gitignore = true
lines_between_types = 1

[tool.poetry]
name = "Luna"
version = "0.1.0"
description = "Mocking project for testing Cosmos"
authors = ["Francesco Milani <fmilani@bink.com>"]

[tool.poetry.dependencies]
python = "^3.9"
falcon = "^3.0.0"
gunicorn = "^20.1.0"
python-dotenv = "^0.17.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.8.0"
flake8 = "^3.9.0"
mypy = "^0.812"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
